A gate for the standalone's uncarried .text, and the warnings it was hiding
The standalone does not carry the original's .text. That range is filled with
int3, so a reference into it is neither a link error nor a crash at the call
site -- it is whatever the filler happens to mean.

That is not hypothetical; it is how this check came to exist. c_dfDIMouse is
a DIDATAFORMAT whose struct sits in .rdata, which IS carried, pointing at an
rgodf array at 0x004643A0, which is NOT. SetDataFormat answered E_INVALIDARG
and the game reported "DDERROR 80070057: SetDataFormat (mouse)" -- a
plausible DirectInput failure with nothing in it about a missing byte range.
A missing free seam did worse and faulted at an address that appears nowhere
in the binary.

tools/checkgap.py states the invariant: any ADDR_ macro src/game
dereferences whose address lies in the gap must be redefined under
AM2_STANDALONE. It is 37 of 37 today, all the statically linked MSVC CRT plus
the stubbed logger, which is exactly the set the port points at the real C
library. Tested by removing one definition, which names it and its use site.

It says what it cannot see: patch_replace targets are skipped, since
installing a patch in the standalone is a no-op; and it resolves macros, not
dataflow, so a seam reached through a variable is invisible -- the blind spot
checkseams.py already documents.

It earned its keep before being committed by disproving something. Movement
not working in the standalone looked like a missing input seam, the mouse
data format having been exactly that. One run said 37 of 37, so it is not a
gap read and the input path did not need reading again.

The gap bound is named once in mkglobals.py rather than repeated at four
sites, and checkgap imports it, so the two cannot disagree about where the
gap is.

Two -Wsign-compare warnings are fixed while they were visible. They are
benign -- the loop counter runs 0 to 3 against an unsigned 4 -- but this
file's own step 0 is that warning noise is what hides the next real one, and
-Wunused-function on a static is how an uninstalled reconstruction announces
itself. Both builds now compile with zero warnings. The change is a cast in a
comparison of two small non-negative values and cannot alter behaviour.

make check runs 58 tools.

+++ src/game/win32/widget.cpp
@@ -10729,7 +10729,7 @@ void __cdecl FreeAimSprites(void)
          spr < (AM2_Sprite **)(uintptr_t)ADDR_AIM_LIVE_A; spr++)
         ReleaseSprite(*spr);
 
-    for (i = 0; i < AM2_AIM_ARMIES; i++) {
+    for (i = 0; i < (int32_t)AM2_AIM_ARMIES; i++) {
         ((int32_t *)(uintptr_t)ADDR_AIM_LIVE_A)[i]  = 0;
         ((int32_t *)(uintptr_t)ADDR_AIM_STAMP_A)[i] = 0;
     }
@@ -10738,7 +10738,7 @@ void __cdecl FreeAimSprites(void)
          spr < (AM2_Sprite **)(uintptr_t)ADDR_AIM_LIVE_B; spr++)
         ReleaseSprite(*spr);
 
-    for (i = 0; i < AM2_AIM_ARMIES; i++) {
+    for (i = 0; i < (int32_t)AM2_AIM_ARMIES; i++) {
         ((int32_t *)(uintptr_t)ADDR_AIM_LIVE_B)[i]  = 0;
         ((int32_t *)(uintptr_t)ADDR_AIM_FRAME_B)[i] = 0;
         ((int32_t *)(uintptr_t)ADDR_AIM_STAMP_B)[i] = 0;
